% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{call_geolocator}
\alias{call_geolocator}
\title{Call geolocator for one address}
\usage{
call_geolocator(street, city, state, zip = NA)
}
\arguments{
\item{street}{A character string indicating a street name and number}

\item{city}{A character string indicating a city}

\item{state}{A two-digit character string with a state postal code}

\item{zip}{A five-digit character string with a postal zip code. Optional parameter.}
}
\value{
A character string representing the Census block of the supplied
address.

importFrom utils URLencode
importFrom httr GET stop_for_status
}
\description{
Call geolocator for one address
}
