% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{call_geolocator_batch}
\alias{call_geolocator_batch}
\title{Batch Geocoder for the Census API}
\usage{
call_geolocator_batch(
  data,
  Street,
  City,
  State,
  ZIP,
  benchmark = "Public_AR_Current",
  vintage = "Current_Current",
  batch_size = 1000
)
}
\description{
Batch Geocoder for the Census API
}
