% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/places.R
\name{filter_place}
\alias{filter_place}
\title{Filter a \code{places} Spatial object for only those places matching the
contents of the \code{place} vector.}
\usage{
filter_place(places, place)
}
\arguments{
\item{places}{object returned from a call to \code{places}}

\item{place}{a vector of full place names. The function performs the
comparison in a case-insensitive manner.}
}
\description{
Filter a \code{places} Spatial object for only those places matching the
contents of the \code{place} vector.
}
\examples{
\dontrun{
places("Maine") \%>\% filter_place("berwick")
}
}
