% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fips_codes.R
\docType{data}
\name{fips_codes}
\alias{fips_codes}
\title{Dataset with FIPS codes for US states and counties}
\format{
A data frame with 3,256 rows and 5 variables
}
\usage{
data(fips_codes)
}
\description{
Built-in dataset for use with the \code{lookup_code} function.
To access the data directly, issue the command \code{data(fips_codes)}.
}
\details{
\itemize{
\item \code{county}: County name, title-case
\item \code{county_code}: County code. (3-digit, 0-padded, character)
\item \code{state}: Upper-case abbreviation of state
\item \code{state_code}: State FIPS code (2-digit, 0-padded, character)
\item \code{state_name}: Title-case name of state
}
}
\note{
Last updated 2023-05-18
}
\keyword{datasets}
