% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/states.R
\name{grep_state}
\alias{grep_state}
\title{Find states matching a term in a \code{state} object}
\usage{
grep_state(states, term)
}
\arguments{
\item{states}{object returned from a call to \code{state}}

\item{term}{equivalent to the \code{pattern} argument of \code{grep}}
}
\description{
This is just shorthand for
\code{grep(term, list_states(states), value=TRUE, ignore.case=TRUE)}
}
\examples{
\dontrun{
states() \%>\% grep_state("north")
}
}

