% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is_tigris}
\alias{is_tigris}
\title{Returns \code{TRUE} if \code{obj} has a \code{tigris} attribute}
\usage{
is_tigris(obj)
}
\arguments{
\item{obj}{R object to test}
}
\value{
\code{TRUE} if \code{obj} was made by this package
}
\description{
It's unlikely that said object was not created by this package
}

