% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/places.R
\name{list_places}
\alias{list_places}
\title{Return a list of all the places in a \code{places} object}
\usage{
list_places(places, sorted = TRUE)
}
\arguments{
\item{places}{object returned from a call to \code{places}}

\item{sorted}{return the list sorted or in the order found in the shapefile?}
}
\description{
Return a list of all the places in a \code{places} object
}
\examples{
\dontrun{
places("Maine") \%>\% list_places()
}
}
