% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/states.R
\name{list_states}
\alias{list_states}
\title{Return a list of all the states in a \code{state} object}
\usage{
list_states(states, sorted = TRUE)
}
\arguments{
\item{states}{object returned from a call to \code{state}}

\item{sorted}{return the list sorted or in the order found in the shapefile?}
}
\description{
Return a list of all the states in a \code{state} object
}
\examples{
\dontrun{
states() \%>\% list_states()
}
}
