% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{load_tiger_doc_template}
\alias{load_tiger_doc_template}
\title{Documentation Template for Functions that Utilize \code{load_tiger}}
\arguments{
\item{...}{arguments to be passed to internal function \code{load_tiger}, which is
not exported. See Additional Arguments.}

\item{year}{the data year; defaults to 2022}
}
\description{
Documentation Template for Functions that Utilize \code{load_tiger}
}
\details{
This documentation is inherited by \code{tigris} functions that wrap
\code{load_tiger}.
Include the following in function documentation:
\verb{@inheritParams load_tiger_doc_template}
\verb{@inheritSection load_tiger_doc_template Additional Arguments}
}
\section{Additional Arguments}{

Additional arguments that can be passed in \code{...} are:
\itemize{
\item \code{class} Desired class of return object: \code{"sf"} (the default) or \code{"sp"}. sp
classes should be considered deprecated as of tigris version 2.0, but legacy
support is still available.
\item \code{progress_bar} If set to \code{FALSE}, do not display download progress bar
(helpful for R Markdown documents). Defaults to \code{TRUE}.
\item \code{keep_zipped_shapefile} If set to \code{TRUE}, do not delete zipped shapefile
(stored in temporary directory or \code{TIGRIS_CACHE_DIR} depending on the
configuration of global option \code{"tigris_use_cache"}). Defaults to \code{FALSE}.
\item \code{refresh} Whether to re-download cached shapefiles (\code{TRUE} or \code{FALSE}). The
default is either \code{FALSE} or the value of global option \code{"tigris_refresh"}
if it is set. Specifying this argument will override the behavior set in
\code{"tigris_refresh"} global option.
\item \code{filter_by} Geometry used to filter the output returned by the function.
Can be an sf OR sfc object, an object of class \code{bbox}, or a length-4 vector
of format \code{c(xmin, ymin, xmax, ymax)} that can be converted to a bbox.
Geometries that intersect the input to \code{filter_by} will be returned.
\item \code{crs} Output coordinate reference system. All TIGER/Shapefiles use a EPSG
4269 coordinate reference system.
\item \code{query} A SQL query passed to \code{\link[sf:st_read]{sf::st_read()}}.
}
}

