% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{lookup_code}
\alias{lookup_code}
\title{Look up state and county codes}
\usage{
lookup_code(state, county = NULL)
}
\arguments{
\item{state}{String representing the state you'd like to look up.  Accepts state names (spelled correctly), e.g. "Texas", or postal codes, e.g. "TX".}

\item{county}{The name of the county you'll like to search for.  The state that the county is located in must be supplied for this to work, as there are multiple counties with the same names across states.}
}
\description{
Function to look up the FIPS codes for states and optionally counties you'd like to load data for.  As the package functions require the codes to return the data correctly, this function makes it easy to find the codes that you need.
}

