% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.transform.R
\name{map_transform}
\alias{map_transform}
\title{Transform your United States shape file by removing or rotating Alaska and Hawaii.}
\usage{
map_transform(mapdat, cont.us = NA, fifty.states = NA,
  original.positions = FALSE)
}
\arguments{
\item{mapdat}{The spatial data frame you would like to transform.
Any of the spatial objects returned by the tigris package are acceptable.}

\item{cont.us}{If set to TRUE it will return a map of only the lower forty-eight states.}

\item{fifty.states}{If set TRUE it will return Alaska and Hawaii, but not any teirritories
or outlying islands.}

\item{original.positions}{If set to FALSE it will re-scale and move Alaska and Hawaii.
Note, it is set to FALSE by default.}
}
\description{
Transform your United States shape file by removing or rotating Alaska and Hawaii.
}

