% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center_of_pop.R
\name{popcenters}
\alias{popcenters}
\title{Download a Centers of Population shapefile into R.}
\usage{
popcenters(
  geography = c("state", "county", "tract", "blockgroup"),
  state = NULL,
  year
)
}
\arguments{
\item{geography}{One of 'state', 'county', 'tract', or 'blockgroup'}

\item{state}{The state for which to download data. For state population centers
and for year 2000 tracts, only national files are available, so \code{state} should
not be provided.}

\item{year}{The year for which  to download data. Only decennial census years
2000, 2010, and 2020 are available.}
}
\description{
Description from the Census Bureau: "The concept of the center of population as used
by the U.S. Census Bureau is that of a balance point. The center of population is
the point at which an imaginary, weightless, rigid, and flat (no elevation effects)
surface representation of the 50 states (or 48 conterminous states for calculations
made prior to 1960) and the District of Columbia would balance if weights of identical
size were placed on it so that each weight represented the location of one person."
For more information, please see the link provided.
}
\examples{
\dontrun{
library(tigris)
library(ggplot2)
library(sf)

ctrs <- popcenters('county', state = 'wa', year = 2020)
counties <- counties(state = 'wa', year = 2020)

ggplot() + geom_sf(data = counties, fill = 'grey') + geom_sf(data = ctrs, color = 'red')
}
}
\seealso{
\url{https://www.census.gov/geographies/reference-files/time-series/geo/centers-population.2000.html}
}
