% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{rbind_tigris}
\alias{rbind_tigris}
\title{Row-bind \code{tigris} Spatial objects}
\usage{
rbind_tigris(...)
}
\arguments{
\item{...}{individual (optionally names) \code{tigris} Spatial objects or a list of them}
}
\value{
one combined Spatial object
}
\description{
Row-bind \code{tigris} Spatial objects
}

