% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tigris-package.R
\name{tigris-exports}
\alias{\%>\%}
\alias{tigris-exports}
\title{tigris exported operators}
\description{
The following functions are imported and then re-exported
from the tigris package to enable use of the magrittr
pipe operator with no additional library calls
}

