% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tigris-package.R
\docType{package}
\name{tigris}
\alias{tigris}
\alias{tigris-package}
\title{Download and use U.S. Census TIGER shapefiles in R}
\description{
Download and use U.S. Census TIGER shapefiles in R
}
\author{
Kyle Walker

Kyle Walker (@kyle_e_walker)
}

