% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{tigris_type}
\alias{tigris_type}
\title{Get the type of \code{tigris} object \code{obj} is}
\usage{
tigris_type(obj)
}
\arguments{
\item{obj}{R object to test}
}
\value{
character vector containing the \code{tigris} type of \code{obj}
        or \code{NA} if \code{obj} is not a code \code{tigris} object
}
\description{
Get the type of \code{tigris} object \code{obj} is
}
