% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_county}
\alias{validate_county}
\title{Check if county is a FIPS code, full name or abbreviation.}
\usage{
validate_county(
  state,
  county,
  allow_null = TRUE,
  require_county = FALSE,
  multiple = FALSE,
  .msg = is_interactive(),
  error_call = caller_env()
)
}
\arguments{
\item{state}{State name, abbreviation, or FIPS code. Not case sensitive. If a
single digits FIPS code or longer county FIPS code is passed, the values
are padded with "0" to subset to allow use as a state value.}

\item{county}{County name, abbreviation, or FIPS code. Not case sensitive.}

\item{allow_null}{If \code{TRUE} (default), allow \code{NULL} input values by invisibly
returning \code{NULL}.}

\item{require_county}{If \code{FALSE} (default), warn instead of error on invalid
input values.}

\item{multiple}{If \code{FALSE} (default), require input to match a single state
or county. If \code{TRUE}, allow multiple values.}

\item{.msg}{If \code{TRUE} (default), show informative messages about FIPS code
usage. Uses \code{is_interactive()} by default.}

\item{error_call}{The execution environment of a currently running function.}
}
\value{
\itemize{
\item \code{NULL} if input is \code{NULL} and \code{allow_null = TRUE} and \code{require_county = FALSE}
\item valid county FIPS code if input is even pseudo-valid (i.e. single digit but w/in range)
\item \code{NULL} if input is not a valid FIPS code and \code{require_county = FALSE}
\item errors if input is not a valid FIPS code and \code{require_county = TRUE} or if \code{county} matched multiple values and \code{multiple = FALSE}
}
}
\description{
\code{validate_county()} checks if an input \code{county} is a FIPS code, name or
abbreviation. If some or all input values are invalid, error if
\code{require_county = TRUE}, return \code{NULL} if all inputs are invalid, or warn and
return valid input values.
}
