% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/native.R
\name{alaska_native_regional_corporations}
\alias{alaska_native_regional_corporations}
\title{Download an Alaska Native Regional Corporation shapefile into R.}
\usage{
alaska_native_regional_corporations()
}
\description{
Download an Alaska Native Regional Corporation shapefile into R.
}

