% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/water.R
\name{area_water}
\alias{area_water}
\title{Download an area water shapefile into R}
\usage{
area_water(state, county, ...)
}
\arguments{
\item{state}{The two-digit FIPS code of the state of the county you'd like to
download the water features for.  Can also be state name or abbreviation
(case-insensitive).}

\item{county}{The three-digit FIPS code of the county you'd like the water
features for}
}
\description{
Download an area water shapefile into R
}
\seealso{
Other water functions: \code{\link{coastline}};
  \code{\link{linear_water}}
}

