% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/enumeration_units.R
\name{block_groups}
\alias{block_groups}
\title{Download a Census block groups shapefile into R, and optionally subset by county}
\usage{
block_groups(state, county = NULL, cb = FALSE, detailed = TRUE, ...)
}
\arguments{
\item{state}{The two-digit FIPS code (string) of the state you want. Can also
be state name or state abbreviation.}

\item{county}{The three-digit FIPS code (string) of the county you'd like to
subset for, or a vector of FIPS codes if you desire multiple counties}

\item{cb}{If cb is set to TRUE, download a generalized (1:500k)
file.  Defaults to FALSE (the most detailed TIGER/Line file)}
}
\description{
Download a Census block groups shapefile into R, and optionally subset by county
}
\seealso{
Other general area functions: \code{\link{blocks}};
  \code{\link{counties}}; \code{\link{places}};
  \code{\link{pumas}}; \code{\link{school_districts}};
  \code{\link{states}}; \code{\link{tracts}};
  \code{\link{zctas}}
}

