% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/enumeration_units.R
\name{blocks}
\alias{blocks}
\title{Download a Census block shapefile into R}
\usage{
blocks(state, county = NULL)
}
\arguments{
\item{state}{The two-digit FIPS code (string) of the state you want.}

\item{county}{The three-digit FIPS code (string) of the county you'd like to subset for, or a vector of FIPS codes
if you desire multiple counties}
}
\description{
This function will download an entire block shapefile for a selected state into R, and optionally subset by county.
\strong{A warning:} Census block shapefiles are often very large, especially for large states - for example, the
block file for Texas is 462MB zipped, and XGB unzipped!  If you have a slow or unreliable internet connection,
or insufficient memory, this may prove burdensome given that you have to first download by state and then subset.
}

