% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/enumeration_units.R
\name{blocks}
\alias{blocks}
\title{Download a Census block shapefile into R}
\usage{
blocks(state, county = NULL, ...)
}
\arguments{
\item{state}{The two-digit FIPS code (string) of the state you want. Can also
be state name or state abbreviation.}

\item{county}{The three-digit FIPS code (string) of the county you'd like to
subset for, or a vector of FIPS codes if you desire multiple counties.
Can also be a county name or vector of names.}

\item{...}{arguments to be passed to the underlying `load_tiger` function, which is not exported.
Options include \code{refresh}, which specifies whether or not to re-download shapefiles
(defaults to \code{FALSE}), and \code{year}, the year for which you'd like to download data
(defaults to 2014).}
}
\description{
This function will download an entire block shapefile for a selected state
into R, and optionally subset by county. \strong{A warning:} Census block
shapefiles are often very large, especially for large states - for example, the
block file for Texas is 462MB zipped!  If you have a slow or unreliable internet
connection, or insufficient memory, this may prove burdensome given that you
have to first download by state and then subset.
}
\seealso{
Other general area functions: \code{\link{block_groups}};
  \code{\link{counties}}; \code{\link{places}};
  \code{\link{pumas}}; \code{\link{school_districts}};
  \code{\link{states}}; \code{\link{tracts}};
  \code{\link{zctas}}
}

