% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metro_areas.R
\name{combined_statistical_areas}
\alias{combined_statistical_areas}
\title{Download a combined statistical areas shapefile into R}
\usage{
combined_statistical_areas(detailed = TRUE, ...)
}
\arguments{
\item{detailed}{If detailed is set to FALSE, download a generalized (1:500k)
cartographic boundary file.  Defaults to TRUE (the most detailed
TIGER/Line file).}
}
\description{
Download a combined statistical areas shapefile into R
}
\seealso{
Other metro area functions: \code{\link{core_based_statistical_areas}};
  \code{\link{metro_divisions}}; \code{\link{urban_areas}}
}

