% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/legislative.R
\name{congressional_districts}
\alias{congressional_districts}
\title{Download a congressional districts shapefile for the 114th Congress into R}
\usage{
congressional_districts(detailed = TRUE)
}
\arguments{
\item{detailed}{If detailed is set to FALSE, download a generalized (1:500k)
cartographic boundary file.  Defaults to TRUE (the most detailed
TIGER/Line file).}
}
\description{
Download a congressional districts shapefile for the 114th Congress into R
}
\seealso{
Other legislative district functions: \code{\link{state_legislative_districts}}
}

