% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/legislative.R
\name{congressional_districts}
\alias{congressional_districts}
\title{Download a congressional districts shapefile for the 114th Congress into R}
\usage{
congressional_districts(cb = FALSE, resolution = "500k", detailed = TRUE,
  ...)
}
\arguments{
\item{cb}{If cb is set to TRUE, download a generalized (1:500k)
cartographic boundary file.  Defaults to FALSE (the most detailed
TIGER/Line file).}

\item{resolution}{The resolution of the cartographic boundary file (if cb == TRUE).
Defaults to '500k'; options include '5m' (1:5 million) and '20m' (1:20 million).}
}
\description{
Download a congressional districts shapefile for the 114th Congress into R
}
\seealso{
Other legislative district functions: \code{\link{state_legislative_districts}}
}

