% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/enumeration_units.R
\name{congressional_districts}
\alias{congressional_districts}
\title{Download a congressional districts shapefile for the 114th Congress into R}
\usage{
congressional_districts()
}
\description{
Download a congressional districts shapefile for the 114th Congress into R
}

