% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metro_areas.R
\name{core_based_statistical_areas}
\alias{core_based_statistical_areas}
\title{Download a core-based statistical area shapefile into R}
\usage{
core_based_statistical_areas(cb = FALSE, resolution = "500k",
  detailed = TRUE, ...)
}
\arguments{
\item{cb}{If cb is set to TRUE, download a generalized (1:500k)
cartographic boundary file.  Defaults to FALSE (the most detailed
TIGER/Line file).}
}
\description{
Download a core-based statistical area shapefile into R
}
\seealso{
Other metro area functions: \code{\link{combined_statistical_areas}};
  \code{\link{metro_divisions}}; \code{\link{urban_areas}}
}

