% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/enumeration_units.R
\name{counties}
\alias{counties}
\title{Download a US Counties shapefile into R, and optionally subset by state}
\usage{
counties(state = NULL)
}
\arguments{
\item{state}{The two-digit FIPS code (string) of the state you want, or a vector of codes if you want multiple states.}
}
\description{
Download a US Counties shapefile into R, and optionally subset by state
}

