% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumeration_units.R
\name{county_subdivisions}
\alias{county_subdivisions}
\title{Download a county subdivision shapefile into R}
\usage{
county_subdivisions(state, county = NULL, cb = FALSE, detailed = TRUE,
  ...)
}
\arguments{
\item{state}{The two-digit FIPS code (string) of the state you want. Can also
be state name or state abbreviation.}

\item{county}{The three-digit FIPS code (string) of the county you'd like to
subset for, or a vector of FIPS codes if you desire multiple counties.
Can also be a county name or vector of names.}

\item{cb}{If cb is set to TRUE, download a generalized (1:500k)
file.  Defaults to FALSE (the most detailed TIGER/Line file)}

\item{detailed}{(deprecated) Setting detailed to FALSE returns a 1:500k cartographic boundary file.
This parameter will be removed in a future release.}

\item{...}{arguments to be passed to the underlying `load_tiger` function, which is not
exported. Options include \code{refresh}, which specifies whether or not to re-download
shapefiles (defaults to \code{FALSE}), and \code{year}, the year for which you'd like to
 download data (defaults to 2015).}
}
\description{
From the US Census Bureau (see link for source, and more information): "County subdivisions
are the primary divisions of counties and their equivalent entities for the reporting of
decennial census data. They include census county divisions, census subareas, minor civil
divisions, and unorganized territories. They may represent legal or statistical entities.
The 2015 TIGER/Line Shapefiles contain a 5-character FIPS code field for county subdivisions
and an 8-character National Standards (GNIS) code."
}
\examples{
\dontrun{
library(tigris)

or <- county_subdivisions('Oregon', c('Linn', 'Benton'))

plot(or)

}
}
\seealso{
\url{http://www2.census.gov/geo/pdfs/maps-data/data/tiger/tgrshp2015/TGRSHP2015_TechDoc.pdf}

Other general area functions: \code{\link{block_groups}},
  \code{\link{blocks}}, \code{\link{counties}},
  \code{\link{places}}, \code{\link{pumas}},
  \code{\link{school_districts}}, \code{\link{states}},
  \code{\link{tracts}}, \code{\link{zctas}}
}

