% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fips_codes.R
\docType{data}
\name{fips_codes}
\alias{fips_codes}
\title{Dataset with FIPS codes for US states and counties}
\format{\preformatted{'data.frame':	3235 obs. of  5 variables:
 $ state      : chr  "AL" "AL" "AL" "AL" ...
 $ state_code : chr  "01" "01" "01" "01" ...
 $ state_name : chr  "Alabama" "Alabama" "Alabama" "Alabama" ...
 $ county_code: chr  "001" "003" "005" "007" ...
 $ county     : chr  "Autauga County" "Baldwin County" "Barbour County" "Bibb County" ...
}}
\usage{
fips_codes
}
\description{
Built-in dataset for use with the \code{lookup_code} function.  To access the data directly, issue the command \code{load(fips_codes)}.
}
\keyword{datasets}

