% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/water.R
\name{linear_water}
\alias{linear_water}
\title{Download an linear water shapefile into R}
\usage{
linear_water(state, county)
}
\arguments{
\item{state}{The two-digit FIPS code of the state of the county you'd like to download the water features for}

\item{county}{The three-digit FIPS code of the county you'd like the water features for}
}
\description{
Download an linear water shapefile into R
}

