% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/water.R
\name{linear_water}
\alias{linear_water}
\title{Download an linear water shapefile into R}
\usage{
linear_water(state, county, ...)
}
\arguments{
\item{state}{The two-digit FIPS code of the state of the county you'd like to
download the water features for. Can also be state name or abbreviation
(case-insensitive).}

\item{county}{The three-digit FIPS code of the county you'd like the water
features for.  Can also be a county name.}

\item{...}{arguments to be passed to the underlying `load_tiger` function, which is not exported.
Options include \code{refresh}, which specifies whether or not to re-download shapefiles
(defaults to \code{FALSE}), and \code{year}, the year for which you'd like to download data
(defaults to 2014).}
}
\description{
Download an linear water shapefile into R
}
\seealso{
Other water functions: \code{\link{area_water}};
  \code{\link{coastline}}
}

