% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metro_areas.R
\name{metro_divisions}
\alias{metro_divisions}
\title{Download a metropolitan divisions shapefile into R.}
\usage{
metro_divisions(...)
}
\arguments{
\item{...}{arguments to be passed to the underlying `load_tiger` function, which is not exported.
Options include \code{refresh}, which specifies whether or not to re-download shapefiles
(defaults to \code{FALSE}), and \code{year}, the year for which you'd like to download data
(defaults to 2014).}
}
\description{
Download a metropolitan divisions shapefile into R.
}
\seealso{
Other metro area functions: \code{\link{combined_statistical_areas}};
  \code{\link{core_based_statistical_areas}};
  \code{\link{urban_areas}}
}

