% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metro_areas.R
\name{metro_divisions}
\alias{metro_divisions}
\title{Download a metropolitan divisions shapefile into R.}
\usage{
metro_divisions()
}
\description{
Download a metropolitan divisions shapefile into R.
}

