% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metro_areas.R
\name{metro_divisions}
\alias{metro_divisions}
\title{Download a metropolitan divisions shapefile into R.}
\usage{
metro_divisions(year = NULL, ...)
}
\arguments{
\item{year}{The data year (defaults to 2016).}

\item{...}{arguments to be passed to the underlying `load_tiger` function, which is not exported.
Options include \code{refresh}, which specifies whether or not to re-download shapefiles
(defaults to \code{FALSE}), and \code{year}, the year for which you'd like to download data
(defaults to 2016).}
}
\description{
Metropolitan divisions are subdivisions of metropolitan areas with population of at least 2.5 million.  Please note:
not all metropolitan areas have metropolitan divisions.
}
\seealso{
\url{http://www2.census.gov/geo/pdfs/maps-data/data/tiger/tgrshp2015/TGRSHP2015_TechDoc.pdf}

Other metro area functions: \code{\link{combined_statistical_areas}},
  \code{\link{core_based_statistical_areas}},
  \code{\link{new_england}}, \code{\link{urban_areas}}
}
