% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/native.R
\name{native_areas}
\alias{native_areas}
\title{Download an American Indian / Alaska Native / Native Hawaiian Areas shapefile into R.}
\usage{
native_areas(detailed = TRUE)
}
\arguments{
\item{detailed}{If detailed is set to FALSE, download a generalized (1:500k) file.  Defaults to TRUE (the most detailed TIGER/Line file)}
}
\description{
Download an American Indian / Alaska Native / Native Hawaiian Areas shapefile into R.
}

