% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/native.R
\name{native_areas}
\alias{native_areas}
\title{Download an American Indian / Alaska Native / Native Hawaiian Areas shapefile into R.}
\usage{
native_areas(cb = FALSE, detailed = TRUE, ...)
}
\arguments{
\item{cb}{If cb is set to TRUE, download a generalized (1:500k)
file.  Defaults to FALSE (the most detailed TIGER/Line file)}

\item{detailed}{(deprecated) Setting detailed to FALSE returns a 1:500k cartographic boundary file.
This parameter will be removed in a future release.}

\item{...}{arguments to be passed to the underlying `load_tiger` function, which is not exported.
Options include \code{refresh}, which specifies whether or not to re-download shapefiles
(defaults to \code{FALSE}), and \code{year}, the year for which you'd like to download data
(defaults to 2015).}
}
\description{
Description from the Census Bureau: "This shapefile contain both legal and statistical
American Indian, Alaska Native, and Native Hawaiian
entities for which the Census Bureau publishes data. The legal entities consist of federally recognized
American Indian reservations and off-reservation trust land areas, state-recognized American Indian
reservations, and Hawaiian home lands (HHLs)." For more information, please see the link provided.
}
\examples{
\dontrun{
library(tigris)
library(ggplot2)
library(ggthemes)

nat <- native_areas(cb = TRUE)

nat_map <- fortify(nat)

gg <- ggplot()
gg <- gg + geom_map(data=nat_map, map=nat_map,
                    aes(x=long, y=lat, map_id=id),
                    color="black", fill="white", size=0.25)
gg <- gg + coord_map(xlim=c(-179.1506, -129.9795),  # alaska
                     ylim=c(51.2097, 71.4410))
gg <- gg + theme_map()
gg
}
}
\seealso{
\url{http://www2.census.gov/geo/pdfs/maps-data/data/tiger/tgrshp2015/TGRSHP2015_TechDoc.pdf}

Other native/tribal geometries functions: \code{\link{alaska_native_regional_corporations}},
  \code{\link{tribal_block_groups}},
  \code{\link{tribal_census_tracts}},
  \code{\link{tribal_subdivisions_national}}
}

