% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/enumeration_units.R
\name{places}
\alias{places}
\title{Download a Census-designated places shapefile into R}
\usage{
places(state)
}
\arguments{
\item{state}{The two-digit FIPS code (string) of the state you'd like to download.}
}
\description{
Download a Census-designated places shapefile into R
}

