% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/transportation.R
\name{primary_roads}
\alias{primary_roads}
\title{Download a national primary roads shapefile into R}
\usage{
primary_roads()
}
\description{
Download a national primary roads shapefile into R
}
\seealso{
Other transportation functions: \code{\link{primary_secondary_roads}};
  \code{\link{rails}}; \code{\link{roads}}
}

