% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/transportation.R
\name{primary_secondary_roads}
\alias{primary_secondary_roads}
\title{Download a primary & secondary roads shapefile into R}
\usage{
primary_secondary_roads(state)
}
\arguments{
\item{state}{The two-digit FIPS code of the state of the county you'd like to download the roads for}
}
\description{
Download a primary & secondary roads shapefile into R
}

