% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/transportation.R
\name{rails}
\alias{rails}
\title{Download a national rails shapefile into R}
\usage{
rails()
}
\description{
Download a national rails shapefile into R
}
\seealso{
Other transportation functions: \code{\link{primary_roads}};
  \code{\link{primary_secondary_roads}};
  \code{\link{roads}}
}

