% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/transportation.R
\name{rails}
\alias{rails}
\title{Download a national rails shapefile into R}
\usage{
rails(...)
}
\arguments{
\item{...}{arguments to be passed to the underlying `load_tiger` function, which is not exported.
Options include \code{refresh}, which specifies whether or not to re-download shapefiles
(defaults to \code{FALSE}), and \code{year}, the year for which you'd like to download data
(defaults to 2014).}
}
\description{
Download a national rails shapefile into R
}
\seealso{
Other transportation functions: \code{\link{primary_roads}};
  \code{\link{primary_secondary_roads}};
  \code{\link{roads}}
}

