% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{rbind_tigris}
\alias{rbind_tigris}
\title{Row-bind \code{tigris} Spatial objects}
\usage{
rbind_tigris(..., call = caller_env())
}
\arguments{
\item{...}{individual (optionally names) \code{tigris} Spatial objects or a list of them}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
one combined Spatial object
}
\description{
If multiple school district types are rbound, coerces to "sdall" and does it
}
\examples{
\dontrun{
library(tigris)

me_ctys <- list_counties("me")
aw <- lapply(me_ctys$county_code[1:3], function(x) {
  area_water("Maine", x)
}) \%>\%
  rbind_tigris()
}
}
