% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/transportation.R
\name{roads}
\alias{roads}
\title{Download a roads shapefile into R}
\usage{
roads(state, county)
}
\arguments{
\item{state}{The two-digit FIPS code of the state of the county you'd like to download the roads for}

\item{county}{The three-digit FIPS code of the county you'd like the roads for}
}
\description{
Download a roads shapefile into R
}

