% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/enumeration_units.R
\name{school_districts}
\alias{school_districts}
\title{Download a unified school district shapefile into R}
\usage{
school_districts(state, ...)
}
\arguments{
\item{state}{The two-digit FIPS code (string) of the state you want. Can also
be state name or state abbreviation.}

\item{...}{arguments to be passed to the underlying `load_tiger` function, which is not exported.
Options include \code{refresh}, which specifies whether or not to re-download shapefiles
(defaults to \code{FALSE}), and \code{year}, the year for which you'd like to download data
(defaults to 2014).}
}
\description{
Download a unified school district shapefile into R
}
\examples{
\dontrun{
library(tigris)
library(leaflet)

schools <- school_districts("Maine")

leaflet(schools) \%>\%
  addProviderTiles("CartoDB.Positron") \%>\%
  addPolygons(fillColor = "white",
              color = "black",
              weight = 0.5)
}
}
\seealso{
Other general area functions: \code{\link{block_groups}};
  \code{\link{blocks}}; \code{\link{counties}};
  \code{\link{places}}; \code{\link{pumas}};
  \code{\link{states}}; \code{\link{tracts}};
  \code{\link{zctas}}
}

