% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/enumeration_units.R
\name{school_districts}
\alias{school_districts}
\title{Download a unified school district shapefile into R}
\usage{
school_districts(state)
}
\arguments{
\item{state}{The two-digit FIPS code (string) of the state you want.}
}
\description{
Download a unified school district shapefile into R
}

