% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_geometry.R
\name{shift_geometry}
\alias{shift_geometry}
\title{Shift and rescale Alaska, Hawaii, and Puerto Rico in a US-wide sf object}
\usage{
shift_geometry(
  input_sf,
  geoid_column = NULL,
  preserve_area = FALSE,
  position = c("below", "outside")
)
}
\arguments{
\item{input_sf}{The input sf dataset}

\item{geoid_column}{The GEOID column of the dataset that contains a state ID.  If used, will speed up
processing and avoid spatial overlay to infer locations.  Defaults to \code{NULL}.}

\item{preserve_area}{If TRUE, the areas of Alaska/Hawaii/Puerto Rico relative to the continental US
will be preserved.  Defaults to FALSE where Alaska is proportionally smaller
and Hawaii/Puerto Rico are proportionally larger.}

\item{position}{One of \code{"below"} (the default) or \code{"outside"}.  If \code{"below"}, Alaska, Hawaii,
and Puerto Rico will be placed below the continental United States.  If \code{"outside"},
features will be moved outside the continental US; Alaska will be northwest of Washington,
Hawaii southwest of California, and Puerto Rico southeast of Florida.}
}
\value{
The input sf object with transformed geometry
}
\description{
Shift and rescale Alaska, Hawaii, and Puerto Rico in a US-wide sf object
}
