% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_geometry.R
\name{shift_geometry}
\alias{shift_geometry}
\title{Shift and rescale Alaska, Hawaii, and Puerto Rico in a US-wide sf object}
\usage{
shift_geometry(
  input_sf,
  geoid_column = NULL,
  preserve_area = FALSE,
  position = c("below", "outside")
)
}
\arguments{
\item{input_sf}{The input sf dataset}

\item{geoid_column}{The GEOID column of the dataset that contains a state ID.  If used, will speed up
processing and avoid spatial overlay to infer locations.  Defaults to \code{NULL}.}

\item{preserve_area}{If TRUE, the areas of Alaska/Hawaii/Puerto Rico relative to the continental US
will be preserved.  Defaults to FALSE where Alaska is proportionally smaller
and Hawaii/Puerto Rico are proportionally larger.}

\item{position}{One of \code{"below"} (the default) or \code{"outside"}.  If \code{"below"}, Alaska, Hawaii,
and Puerto Rico will be placed below the continental United States.  If \code{"outside"},
features will be moved outside the continental US; Alaska will be northwest of Washington,
Hawaii southwest of California, and Puerto Rico southeast of Florida.}
}
\value{
The input sf object with transformed geometry
}
\description{
Shift and rescale Alaska, Hawaii, and Puerto Rico in a US-wide sf object
}
\examples{
\dontrun{

# Shift and rescale AK/HI/PR for thematic mapping
library(tigris)
library(tidycensus)
library(tidyverse)

us_states <- states(cb = TRUE, resolution = "20m") \%>\%
  shift_geometry()

# Shift but preserve area
us_states_eqarea <- states(cb = TRUE, resolution = "20m") \%>\%
  shift_geometry(preserve_area = TRUE)

# Shift and rescale but position AK/HI/PR outside the continental US rather than below
us_states_outside <- states(cb = TRUE, resolution = "20m") \%>\%
  shift_geometry(position = "outside")

# Shift a dataset obtained outside tigris and make a map
income_by_metro <- get_acs(
  geography = "cbsa",
  variables = "B01002_001",
  geometry = TRUE
) \%>\%
  shift_geometry()
ggplot() +
  geom_sf(data = income_by_metro, aes(fill = estimate), color = NA) +
  geom_sf(data = us_states, fill = NA, color = "black", size = 0.1) +
  scale_fill_viridis_c() +
  theme_void(base_size = 16) +
  labs(title = "Median age by CBSA, 2015-2019 ACS",
       fill = "ACS estimate  ",
       caption = "Note: Alaska, Hawaii, and Puerto Rico are shifted and not to scale.") +
  theme(plot.title = element_text(hjust = 0.5))


}
}
