% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/legislative.R
\name{state_legislative_districts}
\alias{state_legislative_districts}
\title{Download a state legislative districts shapefile into R - upper or lower}
\usage{
state_legislative_districts(state, house = "upper", detailed = TRUE)
}
\arguments{
\item{state}{The two-digit FIPS code (string) of the state.}

\item{house}{Specify here whether you want boundaries for the upper or lower house.  Defaults to "upper".}

\item{detailed}{If detailed is set to FALSE, download a generalized (1:500k) cartographic boundary file.  Defaults
to TRUE (the most detailed TIGER/Line file).}
}
\description{
This function allows you to download boundaries for state legislatures into R.  Generally, state legislatures
are comprised of an "upper" house, which is typically referred to as the Senate, and a "lower" house, which is
often (but not exclusively) referred to as the House.  The exception is Nebraska, which has a unicameral
state legislature.
}

