% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/enumeration_units.R
\name{states}
\alias{states}
\title{Download a states shapefile into R}
\usage{
states(detailed = TRUE)
}
\arguments{
\item{detailed}{If detailed is set to FALSE, download a generalized (1:500k) states file.  Defaults to TRUE
(the most detailed TIGER/Line file)}
}
\description{
Download a states shapefile into R
}

