% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/enumeration_units.R
\name{states}
\alias{states}
\title{Download shapefile for all states into R}
\usage{
states(detailed = TRUE)
}
\arguments{
\item{detailed}{If detailed is set to FALSE, download a generalized (1:500k)
states file.  Defaults to TRUE (the most detailed TIGER/Line file)}
}
\description{
Download shapefile for all states into R
}
\examples{
\dontrun{
library(tigris)
library(leaflet)

states <- states(detailed=FALSE)

leaflet(states) \%>\%
  addProviderTiles("CartoDB.Positron") \%>\%
  addPolygons(fillColor = "white",
              color = "black",
              weight = 0.5) \%>\%
  setView(-98.5795, 39.8282, zoom=3)
}
}
\seealso{
Other general area functions: \code{\link{block_groups}};
  \code{\link{blocks}}; \code{\link{counties}};
  \code{\link{places}}; \code{\link{school_districts}};
  \code{\link{tracts}}; \code{\link{zctas}}
}

