% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/states.R
\name{states}
\alias{states}
\title{Download shapefile for all states into R}
\usage{
states(detailed = TRUE)
}
\arguments{
\item{detailed}{If detailed is set to FALSE, download a generalized (1:500k)
states file.  Defaults to TRUE (the most detailed TIGER/Line file)}
}
\description{
States and Equivalent Entities are the primary governmental divisions of the
United States.  In addition to the 50 states, the Census Bureau treats the
District of Columbia, Puerto Rico, American Samoa, the Commonwealth of the
Northern Mariana Islands, Guam, and the U.S. Virgin Islands as the statistical
equivalents of states for the purpose of data presentation.
}
\examples{
\dontrun{
library(tigris)
library(leaflet)

states <- states(detailed=FALSE)

leaflet(states) \%>\%
  addProviderTiles("CartoDB.Positron") \%>\%
  addPolygons(fillColor = "white",
              color = "black",
              weight = 0.5) \%>\%
  setView(-98.5795, 39.8282, zoom=3)
}
}
\seealso{
\url{https://www.census.gov/geo/reference/gtc/gtc_state.html}

Other general area functions: \code{\link{block_groups}};
  \code{\link{blocks}}; \code{\link{counties}};
  \code{\link{places}}; \code{\link{pumas}};
  \code{\link{school_districts}}; \code{\link{tracts}};
  \code{\link{zctas}}
}

