% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tigris-package.R
\docType{package}
\name{tigris}
\alias{tigris}
\alias{tigris-package}
\title{Download and use U.S. Census TIGER shapefiles in R}
\description{
Use option \code{tigris_use_cache} to tell \code{tigris} to cache
Census shapefile downloads. This is \code{TRUE} by default. e.g.
\code{options(tigris_use_cache=TRUE)}
}
\details{
Use option \code{tigris_refresh} to force a refresh of cached \code{tigris}
Shapefiles. e.g. \code{options(tigris_refresh=TRUE)}

Use option \code{tigris_year} to change the year for which you'd like to download data.
e.g. \code{options(tigris_year = 2013)}.  The default year for the package is 2015, the most
recent year for which data are available.
}
\note{
Three \code{options} control behavior of various \code{tigris} functions.
      See \code{Description} for more information.
}
\author{
Kyle Walker

Kyle Walker (@kyle_e_walker)
}

