% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/enumeration_units.R
\name{tracts}
\alias{tracts}
\title{Download a Census tracts shapefile into R, and optionally subset by county}
\usage{
tracts(state, county = NULL, detailed = TRUE, ...)
}
\arguments{
\item{state}{The two-digit FIPS code (string) of the state you want. Can also
be state name or state abbreviation.}

\item{county}{The three-digit FIPS code (string) of the county you'd like to
subset for, or a vector of FIPS codes if you desire multiple counties}

\item{detailed}{If detailed is set to FALSE, download a generalized (1:500k)
tracts file.  Defaults to TRUE (the most detailed TIGER/Line file)}
}
\description{
Census Tracts are small, relatively permanent statistical subdivisions of
a county or equivalent entity that are updated by local participants prior
to each decennial census as part of the Census Bureau's Participant
Statistical Areas Program. The Census Bureau delineates census tracts in
situations where no local participant existed or where state, local, or
tribal governments declined to participate. The primary purpose of census
tracts is to provide a stable set of geographic units for the presentation
of statistical data.
}
\details{
Census tracts generally have a population size between 1,200 and 8,000 people,
with an optimum size of 4,000 people. A census tract usually covers a
contiguous area; however, the spatial size of census tracts varies widely
depending on the density of settlement.Census tract boundaries are
delineated with the intention of being maintained over a long time so that
statistical comparisons can be made from census to census. Census tracts
occasionally are split due to population growth or merged as a result of
substantial population decline.

Census tract boundaries generally follow visible and identifiable features.
They may follow nonvisible legal boundaries, such as minor civil division
(MCD) or incorporated place boundaries in some states and situations, to
allow for census-tract-to-governmental-unit relationships where the
governmental boundaries tend to remain unchanged between censuses.  State and
county boundaries always are census tract boundaries in the standard census
geographic hierarchy. Tribal census tracts are a unique geographic entity
defined within federally recognized American Indian reservations and
off-reservation trust lands and can cross state and county boundaries.
Tribal census tracts may be completely different from the census tracts
and block groups defined by state and county
}
\seealso{
\url{https://www.census.gov/geo/reference/gtc/gtc_ct.html}

Other general area functions: \code{\link{block_groups}};
  \code{\link{blocks}}; \code{\link{counties}};
  \code{\link{places}}; \code{\link{pumas}};
  \code{\link{school_districts}}; \code{\link{states}};
  \code{\link{zctas}}
}

