% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/enumeration_units.R
\name{tracts}
\alias{tracts}
\title{Download a Census tracts shapefile into R, and optionally subset by county}
\usage{
tracts(state, county = NULL)
}
\arguments{
\item{state}{The two-digit FIPS code (string) of the state you want.}

\item{county}{The three-digit FIPS code (string) of the county you'd like to subset for}
}
\description{
Download a Census tracts shapefile into R, and optionally subset by county
}

