% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/native.R
\name{tribal_census_tracts}
\alias{tribal_census_tracts}
\title{Download a Tribal Census tract shapefile into R.}
\usage{
tribal_census_tracts()
}
\description{
Download a Tribal Census tract shapefile into R.
}
\examples{
\dontrun{
library(tigris)
library(leaflet)

trib <- tribal_census_tracts()
leaflet(trib) \%>\%
  addProviderTiles("CartoDB.Positron") \%>\%
  addPolygons(fillColor = "white",
              color = "black",
              weight = 0.5)
}
}
\seealso{
Other native/tribal geometries functions: \code{\link{alaska_native_regional_corporations}};
  \code{\link{native_areas}};
  \code{\link{tribal_block_groups}};
  \code{\link{tribal_subdivisions_national}}
}

