% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/native.R
\name{tribal_census_tracts}
\alias{tribal_census_tracts}
\title{Download a Tribal Census tract shapefile into R.}
\usage{
tribal_census_tracts()
}
\description{
Download a Tribal Census tract shapefile into R.
}

