% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/native.R
\name{tribal_subdivisions_national}
\alias{tribal_subdivisions_national}
\title{Download an American Indian Tribal Subdivision National shapefile into R.}
\usage{
tribal_subdivisions_national()
}
\description{
Download an American Indian Tribal Subdivision National shapefile into R.
}

