% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metro_areas.R
\name{urban_areas}
\alias{urban_areas}
\title{Download an urbanized areas shapefile into R}
\usage{
urban_areas(cb = FALSE, detailed = TRUE, ...)
}
\arguments{
\item{cb}{If cb is set to TRUE, download a generalized (1:500k)
cartographic boundary file.  Defaults to FALSE (the most detailed
TIGER/Line file).}
}
\description{
Download an urbanized areas shapefile into R
}
\seealso{
Other metro area functions: \code{\link{combined_statistical_areas}};
  \code{\link{core_based_statistical_areas}};
  \code{\link{metro_divisions}}
}

