% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/metro_areas.R
\name{urban_areas}
\alias{urban_areas}
\title{Download an urbanized areas shapefile into R}
\usage{
urban_areas(detailed = TRUE)
}
\arguments{
\item{detailed}{If detailed is set to FALSE, download a generalized (1:500k) cartographic boundary file.  Defaults
to TRUE (the most detailed TIGER/Line file).}
}
\description{
Download an urbanized areas shapefile into R
}

