% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_county}
\alias{validate_county}
\title{Check if county is a FIPS code, full name or abbreviation.}
\usage{
validate_county(
  state,
  county,
  allow_null = TRUE,
  require_county = FALSE,
  multiple = FALSE,
  .msg = is_interactive(),
  error_call = caller_env()
)
}
\arguments{
\item{state}{State name, abbreviation, or FIPS code. Not case sensitive. If a
single digits FIPS code or longer county FIPS code is passed, the values
are padded with "0" to subset to allow use as a state value.}

\item{county}{County name, abbreviation, or FIPS code. Not case sensitive.}
}
\value{
\itemize{
\item \code{NULL} if input is \code{NULL} and \code{allow_null = TRUE} and \code{require_county = FALSE}
\item valid county FIPS code if input is even pseudo-valid (i.e. single digit but w/in range)
\item \code{NULL} if input is not a valid FIPS code and \code{require_county = FALSE}
\item errors if input is not a valid FIPS code and \code{require_county = TRUE} or if \code{county} matched multiple values and \code{multiple = FALSE}
}
}
\description{
\code{\link[=validate_county]{validate_county()}} checks if an input \code{county} is a FIPS code, name or
abbreviation. If some or all input values are invalid, error if
\code{require_county = TRUE}, return \code{NULL} if all inputs are invalid, or warn and
return valid input values.
}
\keyword{internal}
