% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_state}
\alias{validate_state}
\title{Check if state is a FIPS code, full name or abbreviation}
\usage{
validate_state(
  state,
  allow_null = TRUE,
  require_state = FALSE,
  multiple = FALSE,
  .msg = is_interactive(),
  error_call = caller_env()
)
}
\arguments{
\item{state}{State name, abbreviation, or FIPS code. Not case sensitive. If a
single digits FIPS code or longer county FIPS code is passed, the values
are padded with "0" to subset to allow use as a state value.}

\item{allow_null}{If \code{TRUE} (default), allow \code{NULL} input values by invisibly
returning \code{NULL}.}

\item{require_state}{If \code{FALSE} (default), warn instead of error on invalid
input values.}

\item{multiple}{If \code{FALSE} (default), require input to match a single state
or county. If \code{TRUE}, allow multiple values.}
}
\value{
\itemize{
\item \code{NULL} if input is \code{NULL} and \code{allow_null = TRUE} and \code{require_state = FALSE}
\item valid state FIPS code if input is even pseudo-valid (i.e. single digit but w/in range)
\item \code{NULL} if input is not a valid FIPS code and \code{require_state = FALSE}
\item errors if input is not a valid FIPS code and \code{require_state = TRUE} or if \code{state} matched multiple values
}
}
\description{
\code{\link[=validate_state]{validate_state()}} checks if an input state values is a FIPS code, full name
or abbreviation. If some or all input values are invalid, error if
\code{require_state = TRUE}, return \code{NULL} if all inputs are invalid, or warn and
return valid input values.
}
\keyword{internal}
