% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legislative.R
\name{voting_districts}
\alias{voting_districts}
\title{Download a voting districts shapefile into R}
\usage{
voting_districts(state = NULL, county = NULL, cb = FALSE, year = 2020, ...)
}
\arguments{
\item{state}{The state for which you'd like to retrieve data.  Can be a state name,
state abbreviation, or FIPS code. When \code{NULL} and combined with
\code{cb = TRUE}, a national dataset of voting districts will be returned.}

\item{county}{The county for which you are requesting data.  Can be a county name or
FIPS code.  If \code{NULL} (the default), data for the entire state will
be returned.}

\item{cb}{If cb is set to TRUE, download a generalized (1:500k)
cartographic boundary file.  Defaults to FALSE (the most detailed
TIGER/Line file).}

\item{year}{the data year; defaults to 2022}

\item{...}{arguments to be passed to internal function \code{load_tiger}, which is not exported. See Additional Arguments.}
}
\description{
Obtain feature geometry for 2020 voting districts, which align with voting districts for the
2020 PL-94171 redistricting data from the US Census Bureau.
}
\details{
The US Census Bureau describes \emph{voting districts} as follows:
Voting district (VTD) is a generic term adopted by the Bureau of the Census
to include the wide variety of small polling areas, such as election districts,
precincts, or wards, that State and local governments create for the purpose
of administering elections. Some States also use groupings of these entities
to define their State and local legislative districts, as well as the districts they
define for election of members to the U.S. House of Representatives. In a
nationwide cooperative program for the 1980 census, the Census Bureau
gave States the opportunity to request use of these election precinct boundaries as the boundaries of
census enumeration districts (EDs) or, in some areas, census blocks.

Support for voting districts in tigris 1.5 and higher is aligned with the 2020 PL redistricting
data.  The argument \code{cb = FALSE} retrieves voting districts from the TIGER/Line PL
shapefiles.  A generalized version from the cartographic boundary dataset is available with the
argument \code{cb = TRUE}.
}
\section{Additional Arguments}{

Additional arguments that can be passed in \code{...} are:
\itemize{
\item \code{class} Desired class of return object: \code{"sf"} (the default) or \code{"sp"}.  sp classes should be considered deprecated as of tigris version 2.0, but legacy support is still available.
\item \code{progress_bar} If set to \code{FALSE}, do not display download progress bar (helpful for R Markdown documents). Defaults to \code{TRUE}.
\item \code{keep_zipped_shapefile} If set to \code{TRUE}, do not delete zipped shapefile (stored in temporary directory or \code{TIGRIS_CACHE_DIR}
depending on the configuration of global option \code{"tigris_use_cache"}). Defaults to \code{FALSE}.
\item \code{refresh} Whether to re-download cached shapefiles (\code{TRUE} or \code{FALSE}) . The default is either \code{FALSE} or the value of global
option \code{"tigris_refresh"} if it is set. Specifying this argument will override the behavior set in \code{"tigris_refresh"} global option.
\item \code{filter_by} Geometry used to filter the output returned by the function.  Can be an sf object, an object of class \code{bbox}, or a length-4 vector of format \code{c(xmin, ymin, xmax, ymax)} that can be converted to a bbox. Geometries that intersect the input to \code{filter_by} will be returned.
}
}

\examples{
\dontrun{#'
library(tigris)

ia <- voting_districts("Iowa")

plot(ia$geometry)

}
}
\seealso{
\url{https://www2.census.gov/geo/pdfs/reference/GARM/Ch14GARM.pdf}

Other legislative district functions: 
\code{\link{congressional_districts}()},
\code{\link{state_legislative_districts}()}
}
\concept{legislative district functions}
