% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/enumeration_units.R
\name{zctas}
\alias{zctas}
\title{Download a Zip Code Tabulation Area (ZCTA) shapefile into R}
\usage{
zctas(detailed = TRUE)
}
\arguments{
\item{detailed}{If detailed is set to FALSE, download a generalized (1:500k) ZCTA file.  Defaults to TRUE
(the most detailed TIGER/Line file)}
}
\description{
Download a Zip Code Tabulation Area (ZCTA) shapefile into R
}

